feat(lib/smmuv3): derive StreamID and SMMU index for TDI identifier
Extend smmuv3_configure_stream() to return the SMMU index and the
derived StreamID for the given TDI identifier, and store these values
in the vdev structure.
The SMMU index and mapping offset are obtained from the Root Complex
list by scanning the BDF mapping ranges to locate the entry
corresponding to the TDI identifier.
The StreamID is then derived on the selected SMMU instance from the
RequesterID (TDI identifier) using the mapping offset, as defined by
Arm Base System Architecture v1.2.

+++ lib/smmuv3/src/aarch64/smmuv3_tlbi.c
@@ -197,7 +197,7 @@ int smmuv3_inv(void)
 	return ret;
 }
 
-int smmuv3_inv_entries(unsigned long smmu_idx, unsigned int vmid)
+int smmuv3_inv_entries(unsigned int smmu_idx, unsigned int vmid)
 {
 	struct smmuv3_dev *smmu;
 	int ret;
